+++ src/common/modals/base.vue
@@ -33,7 +33,7 @@ export default class ModalBase extends Vue {
 </script>
 
 <style scoped lang="scss">
-@import '../../../colors';
+@import '../../colors';
 
 .modal {
   position: absolute;
